add separate files for the scripts and fix some bugs
